backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
anthropic==0.40.0
aiosqlite==0.20.0
python-multipart==0.0.20
reportlab==4.2.5
httpx==0.27.2
pydantic==2.10.3
pdfminer.six==20231228
pymupdf>=1.24.0
